Remove winrt support
