+++ Cargo.toml
@@ -22,12 +22,11 @@ targets = [
 ]
 
 [features]
-default = [ "file-drop", "protocol", "win32", "tray" ]
+default = [ "file-drop", "protocol", "tray" ]
 file-drop = [ ]
 protocol = [ ]
 dox = [ "tao/dox" ]
 tray = [ "tao/tray" ]
-win32 = [ "webview2", "winapi", "webview2-sys" ]
 
 [dependencies]
 libc = "0.2"
@@ -54,11 +53,9 @@ gtk = "0.9"
 gdk = "0.13"
 
 [target."cfg(target_os = \"windows\")".dependencies]
-webview2 = { version = "0.1.1", optional = true }
-webview2-sys = { version = "0.1.0", optional = true }
-winapi = { version = "0.3", features = [ "libloaderapi", "oleidl" ], optional = true }
-windows-webview2 = { version = "0.1", optional = true }
-windows = { version = "0.7", optional = true }
+webview2 = "0.1"
+webview2-sys = "0.1"
+winapi = { version = "0.3", features = [ "libloaderapi", "oleidl" ] }
 
 [target."cfg(any(target_os = \"ios\", target_os = \"macos\"))".dependencies]
 cocoa = "0.24"
