+++ src/lib.rs
@@ -46,7 +46,7 @@
 //! ## Feature flags
 //!
 //! Wry uses a set of feature flags to toggle several advanced features. `file-drop`, `protocol`,
-//! `tray`, and `win32` are enabled by default.
+//! and `tray` are enabled by default.
 //!
 //! - `file-drop`: Enables [`with_file_drop_handler`] to control the behaviour when there are files
 //! interacting with the window.
@@ -56,8 +56,6 @@
 //! those types if you disable it. They just don't create the actual objects. We set this flag
 //! because some implementations require more installed packages. Disable this if you don't want
 //! to install `libappindicator` and `clang` package.
-//! - `win32`: Enables purely Win32 APIs to build the WebView on **Windows**. This makes backward
-//! compatibility down to Windows 7 possible.
 //! - `dox`: Enables this in `package.metadata.docs.rs` section to skip linking some **Linux**
 //! libraries and prevent from building documentation on doc.rs fails.
 //!
@@ -153,11 +151,6 @@ pub enum Error {
   #[error("Icon error: {0}")]
   Icon(#[from] BadIcon),
   #[cfg(target_os = "windows")]
-  #[cfg(feature = "winrt")]
-  #[error(transparent)]
-  WindowsError(#[from] windows::Error),
-  #[cfg(target_os = "windows")]
-  #[cfg(feature = "win32")]
   #[error(transparent)]
   WebView2Error(#[from] webview2::Error),
 }
